Added SSL certificates and keys  SSL will be used throughout the entire application
Basic Devise configuration
  - Added User model and relative migrations
  - Added user's Authorization model and relative migrations
  - Overridden devise's default OmniauthCallbacks controller
  - Copied devise's default views

Added Welcome controller that will function as the application's
entrypoint.
Welcome controller will prompt the user for login via a selection
of omniauth providers.
Once authenticated the user will be redirected to a MainController
which will serve a different layout which will bootstrap AngularJs
application.

The application will expose two kind of routes, of which one will
be scoped and will serve the public API.
The other will be used only for user authentication.